+++ packages/api-reference/src/components/ApiClientModal.vue
@@ -4,7 +4,7 @@ import {
   useAuthenticationStore,
   useServerStore,
 } from '#legacy'
-import type { SpecConfiguration } from '@scalar/oas-utils'
+import type { Spec, SpecConfiguration } from '@scalar/oas-utils'
 import { type App, onBeforeUnmount, onMounted, ref, watch } from 'vue'
 
 import { apiClientBus, modalStateBus } from './api-client-bus'
@@ -12,6 +12,7 @@ import { apiClientBus, modalStateBus } from './api-client-bus'
 const props = defineProps<{
   proxyUrl?: string
   spec?: SpecConfiguration
+  servers?: Spec['servers']
 }>()
 
 const el = ref<HTMLDivElement | null>(null)
@@ -29,6 +30,7 @@ onMounted(async () => {
     await createApiClientModal(el.value, {
       spec: props.spec ?? {},
       proxyUrl: props.proxyUrl,
+      servers: props.servers,
     })
 
   modalStateBus.emit(modalState)