+++ packages/oas-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @scalar/oas-utils
 
+## 0.2.25
+
+### Patch Changes
+
+- 9cc6311: fix: add overload to servers as prop for create client app
+
 ## 0.2.24
 
 ### Patch Changes